Add calculate_connecting_area.master
Add a function to compute the area between a certain point and all consecutive point pairs in a convex shape. If the point lies inside the shape, that area is equal to the area of the convex shape, otherwise it will be bigger than it, so that is a way to put an equality constraint to determine whether a point is within a convex shape or not. See documentation for further instructions.
This function doesn't work with non-convex shapes, but it is still possible to do obstacle avoidance with non-convex shapes! The non-convex shape just has to be cut up into separate convex parts, and then there has to ba an equality constraint for each of those convex subshapes with the presented function, et voilà!