Remove packages
I realised that it's a bit of a waste of space to have the packages of casadi and optispline on here while they are just available on our computers, so I removed them to make the repo a lot lighter. Just paste the code you need to import the libraries in matlab in the README file so you can easily copy/paste it in the matlab command window when you start it up, and you should be good to go. Sorry for the inconvenience!