Begin of solution of c8
Add solution of the exercise to c8. I don't know if we also have to give an example of a new generated trajectory using the invariant, because the question isn't very clear. Right now, the sequence of images is interpreted (literal copy of the example code of the image processing exercises) and then the fitting is done in a modified fitting code. The only difference is that where vectors of size 3 were defined, they are now of size 2 etc., and I derived the formula for the skew symmetric matrix in 2D and implemented that. But all in all, it's extremely little work for 3 points in comparison with the rest...

Also made a small change to the b2a to make it look nicer. For some reason it now works better than before, probably some low level stuff.