Alternative solution to a2 (Huber norm)
Alternative solution to a2 that doesn't use weighting but the piecewise definition of the norm. First, a lookup table is created with the value of the huber norm for a set of grid points, and then the objective interpolates that function for the independent variable smoothly with bspline interpolation. Therefore, the solution isn't exact, but what you could do is first use a coarse grid to find an approximate solution, then redo the calculation with a smaller but finer grid around that solution etc, until the desired accuracy is reached. In short, a solution with arbitrary accuracy is possible in a way that uses the piecewise definition instead of the weighing (I'll still look into the validity of that).